Move Department & Region constants to utils
