+++ lemarche/perimeters/models.py
@@ -7,7 +7,7 @@ from django.db import models
 from django.template.defaultfilters import slugify
 from django.utils import timezone
 
-from lemarche.siaes.constants import DEPARTMENTS_PRETTY, REGIONS
+from lemarche.utils.constants import DEPARTMENTS_PRETTY, REGIONS
 from lemarche.utils.fields import ChoiceArrayField
